% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constraints.R
\name{add_constraints}
\alias{add_constraints}
\title{Add constraints to a neural network}
\usage{
add_constraints(object, type = c("l1_norm", "l2_norm"), ...)
}
\arguments{
\item{object}{A neural network object.}

\item{type}{Constraint type. Currently, \code{l1_norm} and \code{l2_norm} are supported.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{nn2poly} neural network object.
}
\description{
This function sets up a neural network object with the constraints required
by the \code{\link{nn2poly}} algorithm.
}
