% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_string_to_function.R
\name{change_string_to_function}
\alias{change_string_to_function}
\title{Auxiliar function used to transform some of the used activation functions
in the neural network from their string name into an R function.}
\usage{
change_string_to_function(af_string_list)
}
\arguments{
\item{af_string_list}{List of strings containing the predefined possible
activation functions, i.e., "softplus", "tanh", "sigmoid" or "linear".}
}
\value{
List of R functions associated with the string names provided.
}
\description{
Auxiliar function used to transform some of the used activation functions
in the neural network from their string name into an R function.
}
