% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weight_dimensions.R
\name{check_weights_dimensions}
\alias{check_weights_dimensions}
\title{dimensions checker}
\usage{
check_weights_dimensions(weights_matrices_list)
}
\arguments{
\item{weights_matrices_list}{List of weight matrices of a given neural network.}
}
\value{
\code{TRUE} if the dimensiones are correct, \code{FALSE} if not.
}
\description{
dimensions checker
}
