% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constraints.R
\name{constraint_l1_norm}
\alias{constraint_l1_norm}
\title{Custom keras L1 constraint.}
\usage{
constraint_l1_norm(w)
}
\arguments{
\item{w}{weights to be passed to the constraint. In usual keras layers this
would be the kernel weights or the bias weights, but this also works with
the custom layers implemented here, where bias and weights will be joined
together.}
}
\description{
This function implements an L1 constraint to the weights of a NN in keras.
It follows a similar structure as the already implemented L2 norm in keras.
Note that currently it only supports a maximum norm bound equal to 1.
}
\details{
Internal function, not exported.
}
