% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{nn2poly_example0}
\alias{nn2poly_example1}
\alias{nn2poly_example2}
\alias{nn2poly_example3}
\title{Datasets}
\format{
An object of class \code{list} of length 10.

An object of class \code{list} of length 10.

An object of class \code{list} of length 10.

An object of class \code{list} of length 10.
}
\usage{
nn2poly_example0

nn2poly_example1

nn2poly_example2

nn2poly_example3
}
\description{
Datasets
}
\keyword{datasets}
