% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_PR.R
\name{evaluate_PR}
\alias{evaluate_PR}
\title{Evaluates the response value of a polynomual regression}
\usage{
evaluate_PR(x, betas)
}
\arguments{
\item{x}{the input variables}

\item{betas}{the PR coefficients}
}
\value{
a vector with the response Y?
}
\description{
Evaluates the response value of a polynomual regression
}
