% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_PR_surface.R
\name{evaluate_PR_surface}
\alias{evaluate_PR_surface}
\title{Evaluates the response value of a polynomual regression in the explicit case
where p=2 and a surface is egnerated as a matrix.}
\usage{
evaluate_PR_surface(x1, x2, betas)
}
\arguments{
\item{x1}{variable x1}

\item{x2}{variable x2}

\item{betas}{PR coefficients}
}
\value{
matrix response?
}
\description{
Evaluates the response value of a polynomual regression in the explicit case
where p=2 and a surface is egnerated as a matrix.
}
