% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_PR_labels_and_indexes.R
\name{generate_PR_labels_and_indexes}
\alias{generate_PR_labels_and_indexes}
\title{Labels generation for}
\usage{
generate_PR_labels_and_indexes(p, t)
}
\arguments{
\item{p}{number of variables (integer)}

\item{t}{order of the polynomial (integer)}
}
\value{
list with indexes and labels
}
\description{
Generates the labels and indexes used to determine each term of a polynomial
regression for a given number of variables p and order t.
}
