% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_parameters.R
\name{get_model_parameters}
\alias{get_model_parameters}
\title{Auxiliary function. It obtains some parameters of a keras model.}
\usage{
get_model_parameters(model)
}
\arguments{
\item{model}{Keras nn model.}
}
\value{
\code{list} of length 3 where the first dimension is the list of weights,
the second dimension is the list of activation functions as strings and the third
dimension is the number of neurons at each layer.
}
\description{
Auxiliary function. It obtains some parameters of a keras model.
}
