% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_partitions_from_python.R
\name{label_partitions_from_python}
\alias{label_partitions_from_python}
\title{Label partitions opbtained in python according to the polynomial regression
terms.}
\usage{
label_partitions_from_python(all_partitions)
}
\arguments{
\item{all_partitions}{nested lists obtained with python.}
}
\value{
The same nested lists with the corresponding names.
}
\description{
Label partitions opbtained in python according to the polynomial regression
terms.
}
