% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{layers_to_constrain}
\alias{layers_to_constrain}
\title{Get the layers to constraint in a torch model.}
\usage{
layers_to_constrain(model)
}
\arguments{
\item{model}{A torch neural network.}
}
\value{
Vector with the names of the layers to constrain.
}
\description{
Get the layers to constraint in a torch model.
}
