% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{luz_constraint}
\alias{luz_constraint}
\title{Luz constraint callback generator.}
\usage{
luz_constraint(constraint_type = "l1")
}
\arguments{
\item{constraint_type}{The type of the constraint. It should be l1 or l2.
Defaults l1.}
}
\value{
A luz callback that applies the constraint to the torch model.
}
\description{
It generates a callback constraint to use when training a torch neural network
with luz.
}
