% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mreplace}
\alias{mreplace}
\title{Replace multiple characters in a string at the same time}
\usage{
mreplace(str, rep)
}
\arguments{
\item{str}{original string}

\item{rep}{vector containing the needed maping to perform the replacement}
}
\value{
str
}
\description{
Replace multiple characters in a string at the same time
}
