% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mreplace_nn2poly_algorithm.R
\name{mreplace_nn2poly_algorithm}
\alias{mreplace_nn2poly_algorithm}
\title{Computes one or several polynomials to represent a given neural network
using the NN2Poly algorithm.}
\usage{
mreplace_nn2poly_algorithm(
  weights_list,
  af_string_list,
  q_taylor_vector,
  all_partitions,
  historical_coeffs = FALSE
)
}
\arguments{
\item{weights_list}{\code{list} of length L ( number of hidden layers + 1)
containing the weights matrix for each layer.
The expected shape of such matrices at any layer L is of the form
$(h_(l-1) + 1)*(h_l)$, that is, the number of rows is the number of neurons
in the previous layer plus the bias vector, and the number of columns is the
number of neurons in the current layer L. Therefore, each column
corresponds to the weight vector affecting each neuron in that layer.}

\item{af_string_list}{\code{list} of length L containing \code{character}
strings with the names of the activation function used at each layer.
The last element should be "linear" if the output layer has a single linear
output neuron, i.e., when solving a regression problem.}

\item{q_taylor_vector}{\code{vector} of length L containing the degree
(\code{numeric}) up to which Taylor expansion should be performed at each
layer.}

\item{all_partitions}{Optional argument containing the needed multipartitions
as list of lists of lists. If missing, the function computes it first. This
step can be computationally expensive and it is encouraged that the
multipartitions are stored and reused when possible.}

\item{historical_coeffs}{a}
}
\value{
If \code{historical_coeffs = FALSE} (default case), it returns a list
with a vector containing the coefficients of a polynomial associated with
each output unit, i.e., one vector in a regression NN and C vectors in
a classification NN with C possible classes. If
\code{historical_coeffs = TRUE}, it returns a list of length L that for each
layer contains a list with a vector of the coefficients of a polynomial
associated with each unit at that layer. The polynomials obtained at the
hidden layers are not needed to represent the NN but can be used to explore
how the method works.
}
\description{
Performs the full NN2Poly algorithm that obtains polynomial coefficients
for a model that performs closely as a given already trained neural network
using its weights and a Taylor approximation of its activation functions.
}
