% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2pr_algorithm.R
\name{nn2pr_algorithm}
\alias{nn2pr_algorithm}
\title{Performs the full algorithm of obtaining polynomial regression coefficients
from the  weights of a given neural network.}
\usage{
nn2pr_algorithm(weights_list, af_string_list, q_taylor_vector)
}
\arguments{
\item{weights_list}{List containing a matrix of weights and bias for each
layer}

\item{af_string_list}{List with the names of the activation fucntion used
at each layer}

\item{q_taylor_vector}{List containing the degree up to which Taylor
expansion should be performed at each layer.}
}
\value{
list
}
\description{
Performs the full algorithm of obtaining polynomial regression coefficients
from the  weights of a given neural network.
}
