% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_derivatives_list.R
\name{obtain_derivatives_list}
\alias{obtain_derivatives_list}
\title{Title}
\usage{
obtain_derivatives_list(af_string_list, q_taylor_vector)
}
\arguments{
\item{af_string_list}{List with the names of the activation function used
at each layer as a string. Currently accepted values: "softplus", "linear",
"tanh" and "sigmoid.}

\item{q_taylor_vector}{List containing the degree up to which Taylor
expansion should be performed at each layer.}
}
\value{
list of vectors with the derivatives
}
\description{
Title
}
