% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly_algorithm.R
\name{obtain_partitions_with_labels}
\alias{obtain_partitions_with_labels}
\title{This functions will generate the partitions obtained with Knuth's algorithm,
compute their labels and store both things in a list of length 2.}
\usage{
obtain_partitions_with_labels(p, q_max)
}
\arguments{
\item{p}{Number of variables.}

\item{q_max}{Maximum degree of the final polynomial.}
}
\value{
List of length 2 where the first element is a list with the labels
and the second element is a list with the partitions.
}
\description{
This functions will generate the partitions obtained with Knuth's algorithm,
compute their labels and store both things in a list of length 2.
}
\examples{
obtain_partitions_with_labels(2, 3)

}
