% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagonal.R
\name{plot_diagonal}
\alias{plot_diagonal}
\title{Plots a comparison between two sets of points.}
\usage{
plot_diagonal(
  x_axis,
  y_axis,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  plot.line = TRUE
)
}
\arguments{
\item{x_axis}{Values to plot in the \code{x} axis.}

\item{y_axis}{Values to plot in the \code{y} axis.}

\item{xlab}{Lab of the \code{x} axis}

\item{ylab}{Lab of the \code{y} axis.}

\item{title}{Title of the plot.}

\item{plot.line}{If a red line with \code{slope = 1} and \code{intercept = 0} should
be plotted.}
}
\value{
Plot (ggplot object).
}
\description{
If the points come from the predictions of an NN and a PM and the line
\code{(plot.line = TRUE)} is displayed, in case the method does exhibit
asymptotic behavior, the points should not fall in the line.
}
