% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NN_PR_comparison.R
\name{plot_NN_PR_comparison}
\alias{plot_NN_PR_comparison}
\title{Plots a comparison between NN and PR predictions.
If the method doesn't present asymptotic behavior, the point should fall
in the red line.}
\usage{
plot_NN_PR_comparison(prediction_PR, prediction_NN)
}
\arguments{
\item{prediction_PR}{prediction values obtained with the PR}

\item{prediction_NN}{prediction values obtained with the NN}
}
\value{
plot (ggplot object)
}
\description{
Plots a comparison between NN and PR predictions.
If the method doesn't present asymptotic behavior, the point should fall
in the red line.
}
