% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taylor_and_synaptic_potentials.R
\name{plot_taylor_and_synpatic_potentials}
\alias{plot_taylor_and_synpatic_potentials}
\title{Function that allows to take a NN and the data input values
and plot the distribution of data synaptic potentials
(sum of input values * weights) at all neurons together AT EACH HIDDEN LAYER
with the Taylor expansion used in the activation functions.}
\usage{
plot_taylor_and_synpatic_potentials(
  data,
  weights_list,
  af_string_list,
  q_taylor_vector
)
}
\arguments{
\item{data}{a}

\item{weights_list}{a}

\item{af_string_list}{a}

\item{q_taylor_vector}{a}
}
\value{
a
}
\description{
Function that allows to take a NN and the data input values
and plot the distribution of data synaptic potentials
(sum of input values * weights) at all neurons together AT EACH HIDDEN LAYER
with the Taylor expansion used in the activation functions.
}
