% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly.R
\name{predict.nn2poly}
\alias{predict.nn2poly}
\title{Predict method for \code{nn2poly} objects.}
\usage{
\method{predict}{nn2poly}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'nn2poly'.}

\item{newdata}{a}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{matrix} containing the predictions. There is one prediction for
each row in \code{newdata}.
}
\description{
Predicted values obtained with an nn2poly object on some new data.
}
