% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/select_allowed_partitions.R
\name{select_allowed_partitions}
\alias{select_allowed_partitions}
\title{Select the allowed partitions at a given layer computation}
\usage{
select_allowed_partitions(equivalent_label, q_previous_layer, all_partitions)

select_allowed_partitions(equivalent_label, q_previous_layer, all_partitions)
}
\arguments{
\item{q_previous_layer}{a}

\item{all_partitions}{a}

\item{coeff_label}{PR coefficient label}

\item{label}{a}
}
\value{
list


}
\description{
Select the allowed partitions at a given layer computation

Title
}
\examples{


}
