% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{torch_l1_constraint}
\alias{torch_l1_constraint}
\title{l1-constraint}
\usage{
torch_l1_constraint(module)
}
\arguments{
\item{module}{Module from nn_module.}
}
\value{
Module with the l1 constraint applied.
}
\description{
l1-constraint for a module of a torch neural network. Only for internal use.
}
