% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constraints.R
\name{add_constraints}
\alias{add_constraints}
\title{Add custom constraints to the weights of a keras neural network.}
\usage{
add_constraints(model, constraint_type = "l1_norm", ...)
}
\arguments{
\item{model}{A keras model to which the restrictions on the weights are to be applied.}

\item{constraint_type}{The type of constraint you want to apply on the model. Currently,
'l1_norm' and 'l2_norm' can be applied.}

\item{keep_old_weights}{Binary parameters that controls if the weights of
\code{model} are kept in the new constrained network.}
}
\value{
A keras model with the custom constraints applied.
}
\description{
This function is designed to transform a given keras NN (feed forward, dense),
into a NN with the same structure, parameters and weights but with our custom
layers that impose L1 and L2 norm constraints
}
