% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weight_constraints.R
\name{check_weight_constraints}
\alias{check_weight_constraints}
\title{Check weight's vector constraints (including bias)}
\usage{
check_weight_constraints(weights, maxnorm)
}
\arguments{
\item{weights}{List of matrices with the weights of each layer.}

\item{maxnorm}{List of 2 elements: the name of the used norm and its max value.}
}
\value{
List across all layers with a vector containing the norms of each weight vector.
}
\description{
Check weight's vector constraints (including bias)
}
