% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_poly.R
\name{eval_poly}
\alias{eval_poly}
\title{Polynomial evaluation}
\usage{
eval_poly(poly, newdata, monomials = FALSE)
}
\arguments{
\item{poly}{List containing 2 items: \code{labels} and \code{values}.
\itemize{
\item \code{labels}: List of integer vectors with same length (or number of cols)
as \code{values}, where each integer vector denotes the combination of
variables associated to the coefficient value stored at the same position in
\code{values}. That is, the monomials in the polynomial. Note that the
variables are numbered from 1 to p, with the intercept is represented by 0.
\item \code{values}: Matrix (can also be a vector if single polynomial), where
each column represents a polynomial, with same number of rows as the length
of \code{labels}, containing at each row the value of the coefficient
of the monomial given by the equivalent label in that same position.
}

Example: If \code{labels} contains the integer vector \code{c(1,1,3)} at position
5, then the value stored in \code{values} at row 5 is the coefficient
associated with the term \eqn{x_1^2*x_3}.}

\item{newdata}{Input data as matrix, vector or dataframe.
Number of columns (or elements in vector) should be the number of variables
in the polynomial (dimension p). Response variable to be predicted should
not be included.}

\item{monomials}{Boolean determining if the returned item should contain the
evaluations of all the monomials of the provided polynomials
(\code{monomials==TRUE}), or if the final polynomial evaluation should be
computed, i.e., adding up all the monomials (\code{monomials==FALSE}).
Defaults to \code{FALSE}.}
}
\value{
If \code{monomials==FALSE}, returns a matrix containing the
evaluation of the polynomials on the given data. The matrix has dimensions
\code{(n_sample, n_polynomials)}, meaning that each column corresponds to the
result of evaluating all the data for a polynomial. If a single polynomial is
provided, the output is a vector instead of a row matrix.

If \code{monomials==TRUE}, returns a 3D array containing the monomials of
each polynomial evaluated on the given data. The array has dimensions
\code{(n_sample, n_monomial_terms, n_polynomials)}, where element
\code{[i,j,k]} contains the evaluation on observation \code{i} on
monomial \code{j} of polynomial \code{k}, where monomial \code{j} corresponds
to the one on \code{poly$labels[[j]]}.
}
\description{
Evaluates one or several polynomials on the given data.
}
\details{
Note that this function is unstable and subject to change. Therefore it is
not exported but this documentations is left available so users can use it if
needed to simulate data by using \code{nn2poly:::eval_poly()}.
}
\seealso{
\code{eval_poly()} is also used in \code{\link[=predict.nn2poly]{predict.nn2poly()}}.
}
