% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_parameters.R
\name{get_model_parameters}
\alias{get_model_parameters}
\title{Auxiliary function. It obtains some parameters of a keras model.}
\usage{
get_model_parameters(model)
}
\arguments{
\item{model}{Keras nn model.}
}
\value{
\code{list} of length 3 with the following items:
\itemize{
\item \code{weights_list}  list of weights.
\item \code{af_string_list}: the list of activation functions as strings
\item \code{n_nurons}: the number of neurons at each layer.
\item \code{p}: the dimension of the problem, i.e., number of predictor variables.
}
}
\description{
Auxiliary function. It obtains some parameters of a keras model.
}
