% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{luz_model_sequential}
\alias{luz_model_sequential}
\title{Luz Model composed of a linear stack of layers}
\usage{
luz_model_sequential(...)
}
\arguments{
\item{...}{Sequence of modules to be added.}
}
\value{
A \code{nn_sequential} module.
}
\description{
Luz Model composed of a linear stack of layers
}
