% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly.R
\name{nn2poly}
\alias{nn2poly}
\title{nn2poly generic}
\usage{
nn2poly(
  object,
  q_taylor_vector = NULL,
  all_partitions = NULL,
  store_coeffs = FALSE,
  forced_max_Q = NULL,
  ...
)
}
\arguments{
\item{object}{An object for which the computation of the nn2poly algorithm is desired.
For the default case, it should be a \code{list} of length L ( number of hidden layers + 1)
containing the weights matrix for each layer. The name of every element in the list
should be the name of the activation function to be used at each layer.

The expected shape of such matrices at any layer L is of the form
$(h_(l-1) + 1)*(h_l)$, that is, the number of rows is the number of neurons
in the previous layer plus the bias vector, and the number of columns is the
number of neurons in the current layer L. Therefore, each column
corresponds to the weight vector affecting each neuron in that layer.
It could also be a \code{keras.engine.training.Model} model.}

\item{q_taylor_vector}{\code{vector} of length L containing the degree
(\code{numeric}) up to which Taylor expansion should be performed at each
layer.}

\item{all_partitions}{Optional argument containing the needed multipartitions
as list of lists of lists. If \code{NULL}, the function computes it first. This
step can be computationally expensive and it is encouraged that the
multipartitions are stored and reused when possible.}

\item{store_coeffs}{Boolean that determines if all polynomials computed in
the internal layers have to be stored and given in the output (TRUE), or if
only the last layer is needed (FALSE).}

\item{forced_max_Q}{Optional argument: integer that determines the maximum order
that we will force in the final polynomial, discarding terms of higher order
that would naturally arise using all the orders in \code{q_taylor_vector}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{nn2poly}.
If \code{store_coeffs = FALSE} (default case), it returns a list
with an item named \code{labels} that is a list of integer vectors with each the
variables index associated to each polynomial term, and a item named \code{values}
which contains a matrix where each row are the coefficients of the polynomial
associated with an output neuron.

If \code{store_coeffs = TRUE}, it returns a list of length L that for each
layer contains an item as explained before. The polynomials obtained at the
hidden layers are not needed to represent the NN but can be used to explore
how the method works.
}
\description{
nn2poly generic
}
