% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly_algorithm.R
\name{nn2poly_algorithm}
\alias{nn2poly_algorithm}
\title{Computes one or several polynomials to represent a given neural network
using the NN2Poly algorithm.}
\usage{
nn2poly_algorithm(
  weights_list,
  af_string_list,
  q_taylor_vector = NULL,
  all_partitions = NULL,
  store_coeffs = FALSE,
  forced_max_Q = NULL
)
}
\arguments{
\item{weights_list}{\code{list} of length L ( number of hidden layers + 1)
containing the weights matrix for each layer.
The expected shape of such matrices at any layer L is of the form
$(h_(l-1) + 1)*(h_l)$, that is, the number of rows is the number of neurons
in the previous layer plus one (as the bias vector is added in the first row),
and the number of columns is the number of neurons in the current layer L.
Therefore, each column corresponds to the weight vector affecting each neuron
in that layer, from 0 (the bias) in row 1, to neuron h_l in row h_l +1.}

\item{af_string_list}{\code{list} of length L containing \code{character}
strings with the names of the activation function used at each layer.}

\item{q_taylor_vector}{\code{vector} of length L containing the degree
(\code{numeric}) up to which Taylor expansion should be performed at each
layer.}

\item{all_partitions}{Optional argument containing the needed multipartitions
as list of lists of lists. If \code{NULL}, the function computes it first. This
step can be computationally expensive and it is encouraged that the
multipartitions are stored and reused when possible.}

\item{store_coeffs}{Boolean that determines if all polynomials computed in
the internal layers have to be stored and given in the output (TRUE), or if
only the last layer is needed (FALSE). Default is FALSE.}

\item{forced_max_Q}{Optional argument: integer that determines the maximum order
that we will force in the final polynomial, discarding terms of higher order
that would naturally arise using all the orders in \code{q_taylor_vector}.}
}
\value{
If \code{store_coeffs = FALSE} (default case), it returns a list
with an item named \code{labels} that is a list of integer vectors with each the
variables index associated to each polynomial term, and a item named \code{values}
which contains a matrix where each row are the coefficients of the polynomial
associated with an output neuron.

If \code{store_coeffs = TRUE}, it returns a list of length L that for each
layer contains an item as explained before. The polynomials obtained at the
hidden layers are not needed to represent the NN but can be used to explore
how the method works.
}
\description{
Performs the full NN2Poly algorithm that obtains polynomial coefficients
for a model that performs closely as a given already trained neural network
using its weights and a Taylor approximation of its activation functions.
}
