% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly_methods.R
\name{plot.nn2poly}
\alias{plot.nn2poly}
\title{Plot method for \code{nn2poly} objects.}
\usage{
\method{plot}{nn2poly}(x, ..., n = NULL)
}
\arguments{
\item{x}{A \code{nn2poly} object, as returned by the \pkg{nn2poly} algorithm.}

\item{...}{Ignored.}

\item{n}{An integer denoting the number of coefficients to be plotted,
after ordering them by absolute magnitude.}
}
\value{
A plot showing the \code{n} most important coefficients.
}
\description{
A function that takes a polynomial (or several ones) as given by the
\pkg{nn2poly} algorithm, and then plots their absolute magnitude as barplots
to be able to compare the most important coefficients.
}
