% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn2poly.R
\name{predict.nn2poly}
\alias{predict.nn2poly}
\title{S3 method for class 'nn2poly'}
\usage{
\method{predict}{nn2poly}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class inheriting from 'nn2poly'.}

\item{newdata}{Matrix for which predictions are to be made.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix} containing the predictions. There is one prediction for
each row in \code{newdata}.
}
\description{
S3 method for class 'nn2poly'
}
