% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{select_allowed_partitions}
\alias{select_allowed_partitions}
\title{Select the allowed partitions at a given layer computation}
\usage{
select_allowed_partitions(coeff_label, q_previous_layer, all_partitions)
}
\arguments{
\item{coeff_label}{PR coefficient label}

\item{q_previous_layer}{Taylor expansion degree at the previous layer.}

\item{all_partitions}{list of lists of partitions}
}
\value{
list
}
\description{
Select the allowed partitions at a given layer computation
}
