% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_allowed_partitions.R
\name{select_allowed_partitions}
\alias{select_allowed_partitions}
\title{Title}
\usage{
select_allowed_partitions(equivalent_label, q_previous_layer, all_partitions)
}
\arguments{
\item{equivalent_label}{a}

\item{q_previous_layer}{a}

\item{all_partitions}{a}
}
\description{
Title
}
