% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{torch_forward_parser}
\alias{torch_forward_parser}
\title{Parse the forward function of a torch model.}
\usage{
torch_forward_parser(model)
}
\arguments{
\item{model}{A torch neural network.}
}
\value{
A \code{list} with two elements where the first element is a vector with
the name of the functions of the forward function in order and the second element
is the class of those functions in order.
}
\description{
This function assumes that the forward function has been written using the pipe syntax.
}
